+++ libs/backend/modules/tag/src/use-cases/update-tag/tests/update-tag.api.graphql.gen.ts
@@ -1,8 +1,8 @@
 import * as Types from '@codelab/frontend/abstract/codegen'
 
-import { TagFragment } from '../../../../../../../frontend/modules/tag/src/Tag.fragment.graphql.gen'
+import { TagFragment } from '../../../../../../../frontend/modules/tag/src/graphql/Tag.fragment.graphql.gen'
 import { gql } from '@apollo/client'
-import { TagFragmentDoc } from '../../../../../../../frontend/modules/tag/src/Tag.fragment.graphql.gen'
+import { TagFragmentDoc } from '../../../../../../../frontend/modules/tag/src/graphql/Tag.fragment.graphql.gen'
 import * as Apollo from '@apollo/client'
 const defaultOptions = {}
 export type TestUpdateTagMutationVariables = Types.Exact<{
