--- libs/frontend/modules/tag/src/Tag.fragment.graphql
@@ -1,20 +1,0 @@
-fragment TagGraph on TagGraph {
-  vertices {
-    ...Tag
-  }
-  edges {
-    ...TagEdge
-  }
-}
-
-fragment TagEdge on TagEdge {
-  source
-  target
-}
-
-fragment Tag on Tag {
-  id
-  name
-  isRoot
-  children
-}
