+++ libs/frontend/modules/tag/src/graphql/Tag.fragment.graphql
@@ -1,0 +1,6 @@
+fragment Tag on Tag {
+  id
+  name
+  isRoot
+  children
+}
