+++ libs/frontend/modules/tag/src/graphql/TagEdge.fragment.graphql
@@ -1,0 +1,4 @@
+fragment TagEdge on TagEdge {
+  source
+  target
+}
