+++ libs/frontend/modules/tag/src/graphql/TagEdge.fragment.graphql.gen.ts
@@ -1,0 +1,11 @@
+import * as Types from '@codelab/frontend/abstract/codegen'
+
+import { gql } from '@apollo/client'
+export type TagEdgeFragment = { source: string; target: string }
+
+export const TagEdgeFragmentDoc = gql`
+  fragment TagEdge on TagEdge {
+    source
+    target
+  }
+`
