+++ libs/frontend/modules/tag/src/graphql/TagGraph.fragment.graphql
@@ -1,0 +1,8 @@
+fragment TagGraph on TagGraph {
+  vertices {
+    ...Tag
+  }
+  edges {
+    ...TagEdge
+  }
+}
